another updated

+++ src/components/CheckOut/BillCard.jsx
@@ -3,6 +3,7 @@ import React, { useState } from "react";
 import { AiOutlinePlus } from "react-icons/ai";
 import { AiOutlineMinus } from "react-icons/ai";
 import { useDispatch, useSelector } from "react-redux";
+import Swal from "sweetalert2";
 import { RiDeleteBin6Line } from "react-icons/ri";
 import {
   addQuantity,
@@ -20,7 +21,44 @@ const BillCard = ({ card }) => {
     dispatch(decreaseQuantity(card));
   };
   const handleDelete = (card) => {
-    dispatch(cartRemove(card));
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton:
+          "px-4 py-2 bg-green-500 text-white rounded-md ml-4 font-sans",
+        cancelButton: "px-4 py-2 bg-red-500 text-white rounded-md",
+      },
+      buttonsStyling: false,
+    });
+
+    swalWithBootstrapButtons
+      .fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Yes, delete it!",
+        cancelButtonText: "No, cancel!",
+        reverseButtons: true,
+      })
+      .then((result) => {
+        if (result.isConfirmed) {
+          swalWithBootstrapButtons.fire(
+            "Deleted!",
+            "Your file has been deleted.",
+            "success"
+          );
+          dispatch(cartRemove(card));
+        } else if (
+          /* Read more about handling dismissals below */
+          result.dismiss === Swal.DismissReason.cancel
+        ) {
+          swalWithBootstrapButtons.fire(
+            "Cancelled",
+            "Your imaginary file is safe :)",
+            "error"
+          );
+        }
+      });
   };
 
   return (
@@ -39,7 +77,7 @@ const BillCard = ({ card }) => {
       </div>
       <div className="flex justify-between text-black">
         <div>
-          $<span>{card.price * card.cartQuantity}</span>
+          $<span>{(card.price * card.cartQuantity).toFixed(2)}</span>
         </div>
         <div className="flex justify-between items-center">
           <AiOutlineMinus onClick={() => handleMono(card)} />
